Show alumni quote, background, and story
On the home page, if the alumnus has a quote, show it, otherwise use
the background.

On the alumni page, show the background, and only show the links to the
story if the alumnus has one (otherwise no button is shown).

Fix a typo in the background field name (capital B). Adjust all the
alumni JSON files to fix this, and move content to the correct quote,
background, or story field.